backend/api/requirements.txt
flask
mysql-connector-python
python-dotenv
flask-cors
flask-bcrypt
PyJWT
